Move monero ser/deser to monero module
